Move feature class into feature folder.
